Enhance Packing Out For Transit

+++ src/Com.Danliris.Service.Packing.Inventory.Application/ToBeRefactored/DyeingPrintingAreaInput/Shipping/InputShippingProductionOrderViewModel.cs
@@ -33,7 +33,7 @@ namespace Com.Danliris.Service.Packing.Inventory.Application.ToBeRefactored.Dyei
         public string Area { get; set; }
         public string PackingInstruction { get; set; }
         public string Remark { get; set; }
-        public decimal PackingLength { get; set; }
+        public double PackingLength { get; set; }
 
         public int OutputId { get; set; }
